organize settings related components in their own files
